+++ bank/src/main/java/bank/Domain/Bank.java
@@ -1,4 +1,4 @@
 package bank.Domain;
 
-public record Bank(int id, String name) {
+public record Bank(long id, String name) {
 }
