+++ bank/src/main/java/bank/Domain/BankAccount.java
@@ -1,10 +1,10 @@
 package bank.Domain;
 
 public record BankAccount(
-        int id,
-        float balance,
-        int accountHolder,
-        int bankId,
+        long id,
+        long balance,
+        long accountHolder,
+        long bankId,
         CurrencyType currency
 ) {
 }
