+++ bank/src/main/java/bank/Infrastructure/AccountDoesntExistExeption.java
@@ -1,7 +1,7 @@
 package bank.Infrastructure;
 
 public class AccountDoesntExistExeption extends RuntimeException {
-    public AccountDoesntExistExeption(String message, int id) {
+    public AccountDoesntExistExeption(String message, long id) {
         super("Cannot find acount with id "+ id + message) ;
     }    
 }
