+++ bank/src/main/java/bank/Infrastructure/InsufficientFunds.java
@@ -1,7 +1,7 @@
 package bank.Infrastructure;
 
 public class InsufficientFunds extends RuntimeException {
-    public InsufficientFunds(String message, int id) {
+    public InsufficientFunds(String message, long id) {
         super(message + id);
     }
 }
